Split MainViewer into VirtualizedImageViewer and PaginatedItemViewer
